Utility to (re)set accounts for testing
