+++ frontend/src/tests/lib/derived/accounts-list.derived.spec.ts
@@ -1,13 +1,13 @@
 import { nnsAccountsListStore } from "$lib/derived/accounts-list.derived";
-import { icpAccountsStore } from "$lib/stores/icp-accounts.store";
 import { mockMainAccount } from "$tests/mocks/icp-accounts.store.mock";
 import { mockSnsMainAccount } from "$tests/mocks/sns-accounts.mock";
+import { setAccountsForTesting } from "$tests/utils/accounts.test-utils";
 import { get } from "svelte/store";
 
 describe("accounts", () => {
   describe("nnsAccountsListStore", () => {
     it("returns nns accounts in an array", () => {
-      icpAccountsStore.setForTesting({
+      setAccountsForTesting({
         main: mockMainAccount,
         subAccounts: [mockSnsMainAccount],
         hardwareWallets: [],
